Delete application funcionando

+++ 06-Hackathon/Workspace/Acme-AnimalShelter/src/main/java/services/ApplicationService.java
@@ -1,5 +1,6 @@
 package services;
 
+import java.util.ArrayList;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,8 @@ import org.springframework.util.Assert;
 
 import repositories.ApplicationRepository;
 import domain.Application;
+import domain.Pet;
+import domain.Report;
 
 
 @Service
@@ -21,6 +24,15 @@ public class ApplicationService {
 	private ApplicationRepository applicationRepository;
 
 	// Suporting services
+	
+	@Autowired
+	private AdministratorService administratorService;
+	
+	@Autowired
+	private PetService petService;
+	
+	@Autowired
+	private ReportService reportService;
 
 	// Constructors
 
@@ -59,14 +71,35 @@ public class ApplicationService {
 	}
 
 	public void delete(Application application) {
+		this.administratorService.checkAuthority();
+		Assert.isTrue(application.getClosed() == false || application.getClient().isBan());
 		Assert.notNull(application);
 		Assert.isTrue(application.getId() != 0);
 		Assert.isTrue(applicationRepository.exists(application.getId()));
+		
+		//Borramos sus asociaciones
+		Pet pet = application.getPet();
+		pet.setApplication(null);
+		this.petService.save(pet);
+		
+		Report report = application.getReport();
+		this.reportService.delete(report);
+		
 		applicationRepository.delete(application);
 	}
 
 	// Other business methods
 	
+	public Collection<Application> findApplicationsPending(){
+		Collection<Application> applications = new ArrayList<Application>();
+		applications = this.applicationRepository.findApplicationsPending();
+		return applications;
+	}
 	
+	public Collection<Application> findApplicationsClientBan(){
+		Collection<Application> applications = new ArrayList<Application>();
+		applications = this.applicationRepository.findApplicationsClientBan();
+		return applications;
+	}
 
 }
